Init port assembly app notes for RF TX to C
Still WIP

Minor edits:
- slight increase for delay of "wait" loop for better visibility of blinks
- init toggles LED 9 times so remains OFF after successful init

+++ Keil_CC251x/RF/cc_rf_v1.h
@@ -66,7 +66,7 @@
 void rfInit(void); 
 
 // Functions for RF controls
-void rfSend(uint16_t* rfTxBuffer, uint16_t rfTxBufLen);
+void rfSend(uint8_t* rfTxBuffer, uint8_t rfTxBufLen);
 void rfReceive(uint16_t* rfRxBuffer, uint16_t rfRxBufLen); 
 
 // Functions for state control
